feat:1、优化项目架构删除Entity项目统一到Infrastructure中 2、前端按钮功能权限完善 3、现在能正确显示用户信息数据

+++ dotnet/HuanTian.Service/SysService/Dto/SysMenuOutput.cs
@@ -38,7 +38,7 @@ namespace HuanTian.Service
         /// <summary>
         /// 菜单ID
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// 菜单名称
@@ -48,7 +48,7 @@ namespace HuanTian.Service
         /// <summary>
         /// 父菜单ID
         /// </summary>
-        public int ParentId { get; set; }
+        public long ParentId { get; set; }
 
         /// <summary>
         /// 菜单元信息
